+++ apina-core/src/main/kotlin/fi/evident/apina/utils/PropertyUtils.kt
@@ -1,5 +1,10 @@
 package fi.evident.apina.utils
 
+import fi.evident.apina.java.model.JavaMethod
+
+val JavaMethod.propertyName: String
+    get() = propertyNameForGetter(name)
+
 fun propertyNameForGetter(getterName: String): String =
     when {
         getterName.startsWith("get") -> getterName.removePrefix("get").decapitalize()